// src/data/types.ts
export type CandidateStatus =
  | "Draft"
  | "Submitted"
  | "Under Review"
  | "Rejected"
  | "Correction Requested"
  | "HR Manager Approved"
  | "HoD Approved"
  | "Offer Released";